Fixed Set Statement Issues on both Library and Shell.
While fixing this issue, I added more features to DialogueShell, but not
actually has been tested yet. I added verbose feature, What-should-I-do
feature, and finally Read-from-file feature. These things have to be
tested before being pushed into master branch.

+++ DialogueScriptLibrary/AST.cs
@@ -37,7 +37,7 @@ namespace RangHo.DialogueScript
 		/// </summary>
 		/// <param name="value">Value of the AST</param>
 		/// <returns>A Number AST Object</returns>
-		public static AST CreateNumberAST(float value)
+		public static AST CreateNumberAST(int value)
 		{
 			AST result = new AST();
 			result.ASTType = Classification.Number;
